Title: 1까지 나누기, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240623/1까지 나누기/divide-by-1.c
@@ -1,12 +1,12 @@
 #include <stdio.h>
 
 int main() {
-    double n;
+    int n;
     int cnt=0;
     int i=1;
 
-    scanf("%lf" ,&n);
-    double div=n;
+    scanf("%d" ,&n);
+    int div=n;
     while(1)
     {
         if(div<=1)
